Add YAML support for rules config files
Rename rules_toml → rules_config and add YAML parsing alongside
TOML. Format is auto-detected from file extension (.toml, .yml,
.yaml). Both formats share the same serde structs and conversion
logic, so behavior is identical. Includes equivalence test.
